refactoring and RAM loading tweak
